requirements.txt
transformers>=4.30.0
torch>=1.10.0
PyQt6>=6.4.0
pdfplumber>=0.7.6
pdfminer.six>=20220319
pytesseract>=0.3.10
Pillow>=9.0.0
sentence-transformers>=2.2.2
torch>=1.10.0
faiss-cpu>=1.7.2
sumy>=0.11.0
nltk>=3.7
python-docx>=0.8.11
reportlab>=3.6.9
numpy>=1.21.0
lmdb>=1.3.0
lmdbm>=0.98
sumy>=0.11.0
nltk>=3.7